Implemented method to download status avatar

+++ libturpial/api/core.py
@@ -630,13 +630,31 @@ class Core:
             return self.__handle_exception(exc)
 
     def get_profile_image(self, acc_id, user):
+        # Returns the path of profile image in original size
         try:
             account = self.accman.get(str(acc_id))
-            img_content = account.get_profile_image(str(user))
-            img_path = os.path.join(account.config.imgdir, str(user))
-            fd = open(img_path, 'w')
-            fd.write(img_content)
-            fd.close()
+            basename = "%s-%s-profile-image" % (acc_id, user)
+            img_path = os.path.join(account.config.imgdir, basename)
+            if not os.path.isfile(img_path):
+                fd = open(img_path, 'w')
+                fd.write(account.get_profile_image(str(user)))
+                fd.close()
+            return Response(img_path)
+        except Exception, exc:
+            return self.__handle_exception(exc)
+
+    def get_status_avatar(self, status):
+        # Returns the path of profile image for the user who post the status
+        # in avatar size (48x48)
+        try:
+            account = self.accman.get(status.account_id)
+            basename = "%s-%s-avatar-%s" % (status.account_id, status.username, os.path.basename(status.avatar))
+            img_path = os.path.join(account.config.imgdir, basename)
+            if not os.path.isfile(img_path):
+                handle = urllib2.urlopen(status.avatar)
+                fp = open(img_path, 'w')
+                fp.write(handle.read())
+                fp.close()
             return Response(img_path)
         except Exception, exc:
             return self.__handle_exception(exc)
